+++ src/CalcInfSusc.h
@@ -1,13 +1,13 @@
 #ifndef COVIDSIM_CALCINFSUSC_H_
 #define COVIDSIM_CALCINFSUSC_H_
 
-double CalcHouseInf(int, unsigned short int);
-double CalcPlaceInf(int, int, unsigned short int);
-double CalcSpatialInf(int, unsigned short int);
-double CalcPersonInf(int, unsigned short int);
-double CalcHouseSusc(int, unsigned short int, int, int);
-double CalcPlaceSusc(int, int, unsigned short int, int, int);
-double CalcSpatialSusc(int, unsigned short int, int, int);
-double CalcPersonSusc(int, unsigned short int, int, int);
+float CalcHouseInf(int, unsigned short int);
+float CalcPlaceInf(int, int, unsigned short int);
+float CalcSpatialInf(int, unsigned short int);
+float CalcPersonInf(int, unsigned short int);
+float CalcHouseSusc(int, unsigned short int, int, int);
+float CalcPlaceSusc(int, int, unsigned short int, int, int);
+float CalcSpatialSusc(int, unsigned short int, int, int);
+float CalcPersonSusc(int, unsigned short int, int, int);
 
 #endif // COVIDSIM_CALCINFSUSC_
