+++ src/Geometry/Vector2.cpp
@@ -2,16 +2,16 @@
 
 using namespace Geometry;
 
-Vector2d Geometry::operator*(const Vector2d &left, const Vector2f &right){
-	return left * Vector2d(right);
+Vector2f Geometry::operator*(const Vector2f &left, const Vector2d &right){
+	return left * Vector2f(right);
 }
-Vector2d Geometry::operator*(const Vector2f &left, const Vector2d &right){
-	return Vector2d(left) * right;
+Vector2f Geometry::operator*(const Vector2d &left, const Vector2f &right){
+	return Vector2f(left) * right;
 }
 
-Vector2d Geometry::operator-(const Vector2d &left, const Vector2i &right){
-	return left - Vector2d(right);
+Vector2f Geometry::operator-(const Vector2f &left, const Vector2i &right){
+	return left - Vector2f(right);
 }
-Vector2d Geometry::operator-(const Vector2i &left, const Vector2d &right){
-	return Vector2d(left) - right;
+Vector2f Geometry::operator-(const Vector2i &left, const Vector2f &right){
+	return Vector2f(left) - right;
 }
