+++ src/InverseCdf.cpp
@@ -10,11 +10,11 @@
  * @param start_value - the value to use for cdf_values_[CDF_RES]
  * @return void
  */
-void InverseCdf::set_neg_log(double start_value)
+void InverseCdf::set_neg_log(float start_value)
 {
 	cdf_values_[CDF_RES] = start_value;
 	for (int i = 0; i < CDF_RES; i++)
-		cdf_values_[i] = -log(1 - ((double)i) / CDF_RES);
+		cdf_values_[i] = -log(1 - ((float)i) / CDF_RES);
 }
 
 /**
@@ -39,7 +39,7 @@ void InverseCdf::assign_exponent()
  * @param value - value of the exponent
  * @return void
  */
-void InverseCdf::assign_exponent(double value)
+void InverseCdf::assign_exponent(float value)
 {
 	for (int i = 0; i <= CDF_RES; i++)
 	{
@@ -57,14 +57,14 @@ void InverseCdf::assign_exponent(double value)
  * @param timesteps_per_day
  * @return value chosen from InverseCdf data
  */
-unsigned short int InverseCdf::choose(double Mean, int tn, double timesteps_per_day)
+unsigned short int InverseCdf::choose(float Mean, int tn, float timesteps_per_day)
 {
 	unsigned short int Value;
 	int i;
-	double q, ti;
+	float q, ti;
 
 	i = (int)floor(q = ranf_mt(tn) * CDF_RES); //// note q defined here as well as i.
-	q -= ((double)i); //// remainder
+	q -= ((float)i); //// remainder
 
 	//// weighted average (sort of) between quartile values from CDF_RES. logged as it was previously exponentiated in ReadParams. Minus as exp(-cdf) was done in ReadParaams. Sort of
 	ti = -Mean * log(q * cdf_values_[i + 1] + (1.0 - q) * cdf_values_[i]);
