+++ src/Kernels.cpp
@@ -6,7 +6,7 @@
 
 using namespace CovidSim::TBD1;
 
-void KernelLookup::setup(double longest_distance)
+void KernelLookup::setup(float longest_distance)
 {
 	size_t size = (size_t)size_ + 1;
 	lookup_.resize(size);
@@ -14,9 +14,9 @@ void KernelLookup::setup(double longest_distance)
 	delta_ = longest_distance / size_;
 }
 
-void KernelLookup::init(double norm, KernelStruct& kernel)
+void KernelLookup::init(float norm, KernelStruct& kernel)
 {
-	double (KernelStruct::*fp)(double) const;
+	float (KernelStruct::*fp)(float) const;
 
 	if (kernel.type_ == 1)
 		fp = &KernelStruct::exponential;
@@ -65,56 +65,56 @@ void KernelLookup::init(const KernelLookup& lookup, Cell **cell_lookup, int cell
 //// **** //// **** //// **** //// **** //// **** //// **** //// **** //// **** //// **** //// **** //// **** //// **** //// **** //// **** //// **** //// **** //// **** //// ****
 //// **** KERNEL DEFINITIONS
 
-double KernelStruct::exponential(double r2) const
+float KernelStruct::exponential(float r2) const
 {
 	return exp(-sqrt(r2) / scale_);
 }
 
-double KernelStruct::power(double r2) const
+float KernelStruct::power(float r2) const
 {
-	double t = -shape_ * log(sqrt(r2) / scale_ + 1.0);
+	float t = -shape_ * log(sqrt(r2) / scale_ + 1.0);
 	return (t < -690.0) ? 0.0 : exp(t);
 }
 
-double KernelStruct::power_b(double r2) const
+float KernelStruct::power_b(float r2) const
 {
-	double t = 0.5 * shape_ * log(r2 / (scale_ * scale_));
+	float t = 0.5 * shape_ * log(r2 / (scale_ * scale_));
 	return (t > 690.0) ? 0.0 : (1.0 / (exp(t) + 1.0));
 }
 
-double KernelStruct::power_us(double r2) const
+float KernelStruct::power_us(float r2) const
 {
-	double t = log(sqrt(r2) / scale_ + 1.0);
+	float t = log(sqrt(r2) / scale_ + 1.0);
 	return (t < -690.0) ? 0.0 : (exp(-shape_ * t) + p3_ * exp(-p4_ * t)) / (1.0 + p3_);
 }
 
-double KernelStruct::gaussian(double r2) const
+float KernelStruct::gaussian(float r2) const
 {
 	return exp(-r2 / (scale_ * scale_));
 }
 
-double KernelStruct::step(double r2) const
+float KernelStruct::step(float r2) const
 {
 	return (r2 > scale_ * scale_) ? 0.0 : 1.0;
 }
 
-double KernelStruct::power_exp(double r2) const
+float KernelStruct::power_exp(float r2) const
 {
-	double d = sqrt(r2);
-	double t = -shape_ * log(d / scale_ + 1.0);
+	float d = sqrt(r2);
+	float t = -shape_ * log(d / scale_ + 1.0);
 	return (t < -690.0) ? 0.0 : exp(t - pow(d / p3_, p4_));
 }
 
-double KernelLookup::num(double r2) const
+float KernelLookup::num(float r2) const
 {
-	double t = r2 / delta_;
+	float t = r2 / delta_;
 	if (t > size_)
 	{
-		fprintf(stderr, "** %lg  %lg  %lg**\n", r2, delta_, t);
+		fprintf(stderr, "** %g  %g  %g**\n", r2, delta_, t);
 		ERR_CRITICAL("r too large in NumKernel\n");
 	}
 
-	double s = t * expansion_factor_;
+	float s = t * expansion_factor_;
 	if (s < size_)
 	{
 		t = s - floor(s);
